Korrastus
Funktsioonide korrastamine.
