+++ README.md
@@ -1,7 +1,12 @@
-Aine Valikpraktika raames valmiv projekt.  
+Aine Valikpraktika raames valmiv projekt.
 
-Vajalik:  
--npm install firebase  
--npm install react-firebase-hooks  
+Rakenduse testimiseks:
+1. Kopeerida endale repositoorium.
+2. Avada terminal projekti kaustas.
+3. Paigaldada firebase käsuga 'npm install firebase'
+4. Paigaldada firebase-hooks käsuga 'npm install firebase-hooks'
+5. firebase.js.sample faililt eemaldada järelliide .sample.
+6. firebase.js faili kirjutada firebase rakenduse konfiguratsiooni andmed.
+7. Terminalis rakendus käsuga 'npm start' rakendus tööle panna.
 
-firebase.js.sample faili on vaja lisada enda firebase konfiguratsiooni info ning seejärel eemaldada faili nimest liide .sample.
+Moderaatorina rakenduse käivitamiseks on vaja luua endale firebase projekt/rakendus.
