+++ src/containers/EventPage/index.js
@@ -1,23 +1,23 @@
 import React from 'react';
+import { questionAnswerLater, questionAnswer, returnToMain, ModeratorView } from '../../actions/functions';
 import { useState, useEffect } from 'react';
 import db from "../../firebase";
-import { onSnapshot, query, collection, where, orderBy, limit, doc, updateDoc } from '@firebase/firestore';
+import { onSnapshot, query, collection, where, orderBy, limit } from '@firebase/firestore';
 /**
 * @author
 * @function EventPage
 **/
-const roomCode = localStorage.getItem("roomCode");
 var statusCode = localStorage.getItem("statusCode");
 var status;
 // eslint-disable-next-line
 if (statusCode == 1) {
     status = [2];
-// eslint-disable-next-line
-} else if(statusCode == 0){
+    // eslint-disable-next-line
+} else if (statusCode == 0) {
     status = [1, 2];
 }
 
-
+//Konkreetse ürituse leht ürituse korraldajale, kus kuvatakse vastamiseks olevad küsimused, kui ka modereeritavad küsimused.
 export const EventPage = () => {
     const [questions, setQuestions] = useState([]);
     const roomCode = localStorage.getItem("roomCode");
@@ -35,14 +35,14 @@ export const EventPage = () => {
             ),
         [roomCode]
     );
-    if(questions.length < 1) {
+    if (questions.length < 1) {
         return (
             <div>
                 <h1>Siin algab vastamiseks olevate küsimuste ala</h1>
-            <p>Hetkel vastamiseks küsimusi pole</p>
-            <button type="button" onClick={returnToMain}>Tagasi</button>
-            <h1>Siin algab modereerimiseks olevate küsimuste ala</h1>
-            <ModeratorView />
+                <p>Hetkel vastamiseks küsimusi pole</p>
+                <button type="button" onClick={returnToMain}>Tagasi</button>
+                <h1>Siin algab modereerimiseks olevate küsimuste ala</h1>
+                <ModeratorView />
             </div>
         )
     }
@@ -53,71 +53,16 @@ export const EventPage = () => {
 
 
             <div clas="App">
-            <h1>Siin algab vastamiseks olevate küsimuste ala</h1>
+                <h1>Siin algab vastamiseks olevate küsimuste ala</h1>
                 {questions.map(question => (
                     <div id={question.id} value={question.id} key={question.id}>{question.question}<button onClick={() => questionAnswer(question.id)}>Vastatud</button><button onClick={() => questionAnswerLater(question.id)}>Hiljem vastamiseks</button></div>
                 ))}
             </div>
             <div>
-            <h1>Siin algab modereerimiseks olevate küsimuste ala</h1>
-            <ModeratorView />
+                <h1>Siin algab modereerimiseks olevate küsimuste ala</h1>
+                <ModeratorView />
             </div>
         </div>
 
     )
 }
-
-function questionAnswer(questionId) {
-    const roomCode = localStorage.getItem('roomCode');
-    const questionDocRef = doc(db, "events", roomCode, "questions", questionId);
-    const questionDocData = { status: 4 };
-    updateDoc(questionDocRef, questionDocData);
-};
-
-function questionAnswerLater(questionId) {
-    const roomCode = localStorage.getItem('roomCode');
-    const questionDocRef = doc(db, "events", roomCode, "questions", questionId);
-    const questionDocData = { status: 5 };
-    updateDoc(questionDocRef, questionDocData);
-};
-
-function returnToMain() {
-    window.location='/';
-}
-
-function ModeratorView() {
-    const [questions, setQuestions] = useState([]);
-    const roomCode = localStorage.getItem("roomCode");
-    useEffect(
-      () =>
-        onSnapshot(query(
-          collection(db, "events", roomCode, "questions"),
-          //erinevad status olekud: 1. posed, 2. readyforanswer, 3. unsuitable, 4. answered, 5. answerlater
-          where("status", "==", 1),
-          orderBy("created", "asc"),
-          limit(10)),
-          (snapshot) =>
-            setQuestions(snapshot.docs.map((doc) => ({ ...doc.data(), id: doc.id })))
-        ),
-      [roomCode]
-    );
-    return (
-      <div clas="App">
-          {questions.map(question => (
-            <div id={question.id} value={question.id} key={question.id}>{question.question}<button onClick={() => questionApprove(question.id)}>Sobilik</button><button onClick={() => questionInappropriate(question.id)}>Mittesobilik</button></div>
-          ))}
-      </div>
-    );
-  };
-
-  function questionInappropriate(questionId) {
-    const questionDocRef = doc(db, "events", roomCode, "questions", questionId);
-    const questionDocData = { status: 3};
-    updateDoc(questionDocRef, questionDocData);
-  };
-  
-  function questionApprove(questionId) {
-    const questionDocRef = doc(db, "events", roomCode, "questions", questionId);
-    const questionDocData = { status: 2};
-    updateDoc(questionDocRef, questionDocData);
-  };
